Minor refactor and updated libftprintf

+++ srcs/pf_cputchar.c
@@ -6,13 +6,13 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/09 16:05:18 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/06/21 00:05:31 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/06/21 08:43:27 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	pf_cputchar(char c, int *inc)
+void	pf_cputchar(char c, size_t *inc)
 {
 	write(1, &c, 1);
 	++(*inc);
